New GUI widget QgsTextSettingsWidget for text settings, switch labelling dialog to use new widget
